requirements.txt
beautifulsoup4
fastapi
httpx
python-dotenv
uvicorn[standard]
youtube-transcript-api
yt-dlp
